# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api MiniBlog APP Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "mohamed.amine.aloui@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/user/comments/post_id={post_id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get List Of Comments by post id",
                "description": "Returns all comment By post Id",
                "operationId": "getCommentsByPostId",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/user/comments": {
            "post": {
                "tags": [
                    "addComment"
                ],
                "summary": "Adding Comment",
                "description": "Adding a new Comment",
                "operationId": "addingNewComment",
                "responses": {
                    "201": {
                        "description": "Comment Added Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "200": {
                        "description": "Validation Errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/user/comments/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get Comment by ID",
                "description": "Returns Comment by its id",
                "operationId": "getCommentByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete Comment",
                "description": "Delete Comment By ID",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}